refactor: move appearance in handler dir
